refactor: moved employee mocks to employee folder inside mocks folder
